requirements.txt
# install these in your project virtualenv
APScheduler==3.6.3
astroid==2.4.2
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
ciscoconfparse==1.5.29
colorama==0.4.4
cryptography==3.3.1
decorator==4.4.2
dnspython==2.1.0
future==0.18.2
idna==2.10
isort==5.7.0
Jinja2==2.11.2
junos-eznc==2.5.4
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.2
ntc-templates==1.6.0
paramiko==2.7.2
passlib==1.7.4
pycparser==2.20
pyeapi==0.8.4
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
python-telegram-bot==13.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
scp==0.13.3
six==1.15.0
tenacity==6.3.1
textfsm==1.1.0
toml==0.10.2
tornado==6.1
transitions==0.8.6
typed-ast==1.4.2
tzlocal==2.1
urllib3==1.26.3
wrapt==1.12.1
yamlordereddictloader==0.4.0
